test: move to tests folder
